+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.6.9 Under development
 
-- Enh #46: Add helper `AttributeBag` class to centralize attribute bag `add()`, `get()`, `merge()`, `remove()`, and `set()` operations with unit tests (@terabytesoftw)
+- Enh #46: Add helper `AttributeBag` class to centralize `add()`, `get()`, `merge()`, `remove()`, and `set()` operations with unit tests (@terabytesoftw)
 
 ## 0.6.8 February 11, 2026